refactor : Move directory of refresh token
