+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/security/refresh/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.dasd412.api.writerservice.adapter.in.security;
+package com.dasd412.api.writerservice.application.service.security.refresh;
 
 import lombok.AccessLevel;
 import lombok.Getter;
